third_party/WebKit/LayoutTests/fast/canvas/gradient-addColorStop-with-invalid-offset.html

<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
test(function() {

    var gradient = document.createElement('canvas').getContext('2d').createLinearGradient(0, 0, 150, 0);

    try {
      gradient.addColorStop(-1, '#000');
    } catch (ex) {
      assert_equals(ex.name, "IndexSizeError");
    }
    try {
      gradient.addColorStop(2, '#000');
    } catch (ex) {
      assert_equals(ex.name, "IndexSizeError");
    }
    try {
      gradient.addColorStop(Infinity, '#000');
    } catch (ex) {
      assert_equals(ex.name, "TypeError");
    }
    try {
      gradient.addColorStop(-Infinity, '#000');
    } catch (ex) {
      assert_equals(ex.name, "TypeError");
    }
    try {
      gradient.addColorStop(NaN, '#000');
    } catch (ex) {
      assert_equals(ex.name, "TypeError");
    }

}, 'This test checks invalid offsets on gradients.');
</script>
